refactor: @share 폴더를 공용으로 만들기
